Cost is now mapped to the current cost period
Internal cost of a person has been changed into a temporal data structure, where cost can change over time. The cost on the person now maps to the current cost period, so when loading the person, the cost is taken from the the current cost period for that person. When updating the cost on the person, the cost is updated on the current cost period for that person.